+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/instrumentation/SpanAddingClassVisitorFactory.kt
@@ -13,6 +13,7 @@ import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.Optional
 import org.objectweb.asm.ClassVisitor
+import org.slf4j.LoggerFactory
 
 @Suppress("UnstableApiUsage")
 abstract class SpanAddingClassVisitorFactory :
@@ -42,6 +43,10 @@ abstract class SpanAddingClassVisitorFactory :
             AndroidXSQLiteStatement(),
             AndroidXRoomDao()
         )
+
+        private val logger by lazy {
+            LoggerFactory.getLogger(SpanAddingClassVisitorFactory::class.java)
+        }
     }
 
     override fun createClassVisitor(
@@ -55,9 +60,14 @@ abstract class SpanAddingClassVisitorFactory :
                 nextClassVisitor,
                 parameters = parameters.get()
             )
-            ?: error(
-                "${classContext.currentClassData.className} is not supported for instrumentation"
-            )
+            ?: nextClassVisitor.also {
+                logger.warn(
+                    """
+                    ${classContext.currentClassData.className} is not supported for instrumentation.
+                    This is likely a bug, please file an issue at https://github.com/getsentry/sentry-android-gradle-plugin/issues
+                    """.trimIndent()
+                )
+            }
 
     override fun isInstrumentable(classData: ClassData): Boolean =
         instrumentables.any { it.isInstrumentable(classData.toClassContext()) }