+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 3.0.0-alpha.2
 
 * Fix: Do not throw exceptions in case something goes wrong with instrumentation (#217)
 * Add support for dry-run on upload native symbols (#209)